refactor(api): rename reset organization learners password usecase
